monitoring: allow compilation for non-Linux
This commit extracts the Linux-specific code into files that are guarded
by Golang build flags, allowing binaries that import the e2e/monitoring
package to compile on non-Linux operating systems. I tested this with
the following `main.go` and was able to compile a binary for both Darwin
and Windows:
```golang
package main

import (
	"fmt"

	m "github.com/efficientgo/e2e/monitoring"
)

func main() {
	s := m.Service{}
	fmt.Println(s)
}
```
